# Caching GetDescription for improved performance, memory allocation.