--- a/QD.Components.Semantic.Benchmarks/Program.cs
+++ b/QD.Components.Semantic.Benchmarks/Program.cs
@@ -1,0 +1,10 @@
+﻿using BenchmarkDotNet.Running;
+
+namespace QD.Components.Semantic.Benchmarks
+{
+    public class Program
+    {
+		public static void Main(string[] args) =>
+			BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
+	}
+}