--- a/QD.Components.Semantic.Benchmarks/QD.Components.Semantic.Benchmarks.csproj
+++ b/QD.Components.Semantic.Benchmarks/QD.Components.Semantic.Benchmarks.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.12.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\QD.Components.Semantic\QD.Components.Semantic.csproj" />
+  </ItemGroup>
+
+</Project>