--- a/QD.Components.Semantic.sln
+++ b/QD.Components.Semantic.sln
@@ -12,6 +12,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Editor", "Editor", "{3D5FB9
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QD.Components.Semantic.Benchmarks", "QD.Components.Semantic.Benchmarks\QD.Components.Semantic.Benchmarks.csproj", "{99B55D7F-62A7-4C7F-972A-E12C0C35DEBE}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -26,6 +28,10 @@ Global
 		{2049C9EA-14B7-41F6-BE3F-F370C12857BB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2049C9EA-14B7-41F6-BE3F-F370C12857BB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2049C9EA-14B7-41F6-BE3F-F370C12857BB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{99B55D7F-62A7-4C7F-972A-E12C0C35DEBE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{99B55D7F-62A7-4C7F-972A-E12C0C35DEBE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{99B55D7F-62A7-4C7F-972A-E12C0C35DEBE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{99B55D7F-62A7-4C7F-972A-E12C0C35DEBE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE