This PR adds caching to GetDescription, as well as a benchmark project that compares the cached and uncached versions in terms of performance and memory allocation. On my machine:

|   Method |        Mean |     Error |    StdDev | Ratio |    Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |------------:|----------:|----------:|------:|---------:|------:|------:|----------:|
| Original | 6,769.37 us | 37.350 us | 29.160 us |  1.00 | 242.1875 |     - |     - |  758.3 KB |
|   Cached |    91.18 us |  3.545 us |  3.143 us |  0.01 |  37.5977 |     - |     - |  115.5 KB |