@Daniel127 2019-12-30
Thanks for the PR, I hope to finish first version of the package in January.

Happy holidays!.